Initialize FrozenBatchNorm2D with num_batches_tracked
Summary: When using D2Go to freeze layers with `FrozenBatchNorm2D`, the function `FrozenBatchNorm2D.convert_frozen_batchnorm` is utilized to replace the original `BatchNorm`. However, during conversion, the attribute `num_batches_tracked` is not copied, which can lead to issues when the model is unfrozen during TorchScript tracking due to conditionals in `FrozenBatchNorm2D.forward`. To avoid these problems, it is necessary to include a copy of the attribute `num_batches_tracked` in `FrozenBatchNorm2D.convert_frozen_batchnorm`, as presented in this diff.

Reviewed By: FarooqZuberi

Differential Revision: D45593607

fbshipit-source-id: 3a119e2741b80c24cdb24a637491c6c64e5dfeb0